refactor: split agui/utils.py into focused modules
utils.py grew to 773 lines mixing state tracking, message extraction,
per-chunk Agno→AG-UI translation, end-of-stream cleanup, and the
top-level streaming loops. Split into 5 focused modules:

  state.py       EventBuffer + validate_agui_state + state delta
  messages.py    text extraction from AG-UI messages / chunks
  events.py      per-chunk Agno → AG-UI dispatch
  completion.py  end-of-stream cleanup (RUN_FINISHED, snapshot)
  streaming.py   sync + async top-level loops

Pure import-graph refactor — no behavior change. router.py and the
two AG-UI test files updated to import from the new module paths.
All 57 AG-UI unit tests pass.
