+++ libs/agno/agno/os/interfaces/agui/completion.py
@@ -1,0 +1,131 @@
+"""End-of-stream cleanup events for AG-UI streams (orphan closure, final snapshot, RUN_FINISHED)."""
+
+import copy
+import json
+import uuid
+from typing import Any, Dict, List, Optional, Union
+
+from ag_ui.core import (
+    BaseEvent,
+    EventType,
+    ReasoningEndEvent,
+    ReasoningMessageEndEvent,
+    RunFinishedEvent,
+    StateSnapshotEvent,
+    TextMessageContentEvent,
+    TextMessageEndEvent,
+    TextMessageStartEvent,
+    ToolCallArgsEvent,
+    ToolCallEndEvent,
+    ToolCallStartEvent,
+)
+
+from agno.os.interfaces.agui.state import EventBuffer
+from agno.run.agent import RunOutputEvent, RunPausedEvent
+from agno.run.team import TeamRunOutputEvent
+
+
+def create_completion_events(
+    chunk: Union[RunOutputEvent, TeamRunOutputEvent],
+    event_buffer: EventBuffer,
+    message_started: bool,
+    message_id: str,
+    thread_id: str,
+    run_id: str,
+    run_state: Optional[Dict[str, Any]] = None,
+) -> List[BaseEvent]:
+    """Create events for run completion."""
+    events_to_emit: List[BaseEvent] = []
+
+    # Close orphaned reasoning session if stream ended mid-reasoning
+    if event_buffer.reasoning_message_id is not None:
+        events_to_emit.append(
+            ReasoningMessageEndEvent(type=EventType.REASONING_MESSAGE_END, message_id=event_buffer.reasoning_message_id)
+        )
+        events_to_emit.append(
+            ReasoningEndEvent(type=EventType.REASONING_END, message_id=event_buffer.reasoning_message_id)
+        )
+        event_buffer.end_reasoning()
+
+    # End remaining active tool calls if needed
+    for tool_call_id in list(event_buffer.active_tool_call_ids):
+        if tool_call_id not in event_buffer.ended_tool_call_ids:
+            end_event = ToolCallEndEvent(
+                type=EventType.TOOL_CALL_END,
+                tool_call_id=tool_call_id,
+            )
+            events_to_emit.append(end_event)
+
+    # End the message and run, denoting the end of the session
+    if message_started:
+        end_message_event = TextMessageEndEvent(type=EventType.TEXT_MESSAGE_END, message_id=message_id)
+        events_to_emit.append(end_message_event)
+
+    # Emit external execution tools
+    if isinstance(chunk, RunPausedEvent):
+        external_tools = chunk.tools_awaiting_external_execution
+        if external_tools:
+            # First, emit an assistant message for external tool calls
+            assistant_message_id = str(uuid.uuid4())
+            assistant_start_event = TextMessageStartEvent(
+                type=EventType.TEXT_MESSAGE_START,
+                message_id=assistant_message_id,
+                role="assistant",
+            )
+            events_to_emit.append(assistant_start_event)
+
+            # Add any text content if present for the assistant message
+            if chunk.content:
+                content_event = TextMessageContentEvent(
+                    type=EventType.TEXT_MESSAGE_CONTENT,
+                    message_id=assistant_message_id,
+                    delta=str(chunk.content),
+                )
+                events_to_emit.append(content_event)
+
+            # End the assistant message
+            assistant_end_event = TextMessageEndEvent(
+                type=EventType.TEXT_MESSAGE_END,
+                message_id=assistant_message_id,
+            )
+            events_to_emit.append(assistant_end_event)
+
+            # Emit tool call events for external execution
+            for tool in external_tools:
+                if tool.tool_call_id is None or tool.tool_name is None:
+                    continue
+
+                start_event = ToolCallStartEvent(
+                    type=EventType.TOOL_CALL_START,
+                    tool_call_id=tool.tool_call_id,
+                    tool_call_name=tool.tool_name,
+                    parent_message_id=assistant_message_id,  # Use the assistant message as parent
+                )
+                events_to_emit.append(start_event)
+
+                args_event = ToolCallArgsEvent(
+                    type=EventType.TOOL_CALL_ARGS,
+                    tool_call_id=tool.tool_call_id,
+                    delta=json.dumps(tool.tool_args),
+                )
+                events_to_emit.append(args_event)
+
+                end_event = ToolCallEndEvent(
+                    type=EventType.TOOL_CALL_END,
+                    tool_call_id=tool.tool_call_id,
+                )
+                events_to_emit.append(end_event)
+
+    # Emit final state snapshot before finishing the run (only if frontend opted into state tracking)
+    if run_state is not None:
+        # Use session_state from RunCompletedEvent (authoritative) if available, otherwise fall back to run_state.
+        # Deep-copy so the emitted event doesn't alias the live agent state (consistent with set_state_snapshot).
+        authoritative_state = getattr(chunk, "session_state", None)
+        final_state = authoritative_state if authoritative_state is not None else run_state
+        snapshot_event = StateSnapshotEvent(type=EventType.STATE_SNAPSHOT, snapshot=copy.deepcopy(final_state))
+        events_to_emit.append(snapshot_event)
+
+    run_finished_event = RunFinishedEvent(type=EventType.RUN_FINISHED, thread_id=thread_id, run_id=run_id)
+    events_to_emit.append(run_finished_event)
+
+    return events_to_emit
