+++ libs/agno/agno/os/interfaces/agui/streaming.py
@@ -1,0 +1,138 @@
+"""Top-level sync and async streaming loops mapping Agno responses to AG-UI events."""
+
+from collections.abc import Iterator
+from typing import Any, AsyncIterator, Dict, Optional, Union
+
+from ag_ui.core import BaseEvent
+
+from agno.os.interfaces.agui.completion import create_completion_events
+from agno.os.interfaces.agui.events import create_events_from_chunk, emit_event_logic
+from agno.os.interfaces.agui.state import EventBuffer
+from agno.run.agent import RunCompletedEvent, RunEvent, RunOutputEvent
+from agno.run.team import TeamRunEvent, TeamRunOutputEvent
+
+
+def stream_agno_response_as_agui_events(
+    response_stream: Iterator[Union[RunOutputEvent, TeamRunOutputEvent]],
+    thread_id: str,
+    run_id: str,
+    run_state: Optional[Dict[str, Any]] = None,
+) -> Iterator[BaseEvent]:
+    """Map the Agno response stream to AG-UI format, handling event ordering constraints."""
+    message_id = ""  # Will be set by EventBuffer when text message starts
+    message_started = False
+    event_buffer = EventBuffer()
+    stream_completed = False
+
+    # Establish baseline state snapshot for delta tracking
+    if run_state is not None:
+        event_buffer.set_state_snapshot(run_state)
+
+    completion_chunk = None
+
+    for chunk in response_stream:
+        # Check if this is a completion event
+        if (
+            chunk.event == RunEvent.run_completed
+            or chunk.event == TeamRunEvent.run_completed
+            or chunk.event == RunEvent.run_paused
+            or chunk.event == TeamRunEvent.run_paused
+        ):
+            # Store completion chunk but don't process it yet
+            completion_chunk = chunk
+            stream_completed = True
+        else:
+            # Process regular chunk immediately
+            events_from_chunk, message_started, message_id = create_events_from_chunk(
+                chunk, message_id, message_started, event_buffer, run_state=run_state
+            )
+
+            for event in events_from_chunk:
+                events_to_emit = emit_event_logic(event_buffer=event_buffer, event=event)
+                for emit_event in events_to_emit:
+                    yield emit_event
+
+    # Process ONLY completion cleanup events, not content from completion chunk
+    if completion_chunk:
+        completion_events = create_completion_events(
+            completion_chunk, event_buffer, message_started, message_id, thread_id, run_id, run_state=run_state
+        )
+        for event in completion_events:
+            events_to_emit = emit_event_logic(event_buffer=event_buffer, event=event)
+            for emit_event in events_to_emit:
+                yield emit_event
+
+    # Ensure completion events are always emitted even when stream ends naturally
+    if not stream_completed:
+        # Synthesize a completion event so the client always sees RUN_FINISHED
+        synthetic_completion = RunCompletedEvent()
+        completion_events = create_completion_events(
+            synthetic_completion, event_buffer, message_started, message_id, thread_id, run_id, run_state=run_state
+        )
+        for event in completion_events:
+            events_to_emit = emit_event_logic(event_buffer=event_buffer, event=event)
+            for emit_event in events_to_emit:
+                yield emit_event
+
+
+async def async_stream_agno_response_as_agui_events(
+    response_stream: AsyncIterator[Union[RunOutputEvent, TeamRunOutputEvent]],
+    thread_id: str,
+    run_id: str,
+    run_state: Optional[Dict[str, Any]] = None,
+) -> AsyncIterator[BaseEvent]:
+    """Map the Agno response stream to AG-UI format, handling event ordering constraints."""
+    message_id = ""  # Will be set by EventBuffer when text message starts
+    message_started = False
+    event_buffer = EventBuffer()
+    stream_completed = False
+
+    # Establish baseline state snapshot for delta tracking
+    if run_state is not None:
+        event_buffer.set_state_snapshot(run_state)
+
+    completion_chunk = None
+
+    async for chunk in response_stream:
+        # Check if this is a completion event
+        if (
+            chunk.event == RunEvent.run_completed
+            or chunk.event == TeamRunEvent.run_completed
+            or chunk.event == RunEvent.run_paused
+            or chunk.event == TeamRunEvent.run_paused
+        ):
+            # Store completion chunk but don't process it yet
+            completion_chunk = chunk
+            stream_completed = True
+        else:
+            # Process regular chunk immediately
+            events_from_chunk, message_started, message_id = create_events_from_chunk(
+                chunk, message_id, message_started, event_buffer, run_state=run_state
+            )
+
+            for event in events_from_chunk:
+                events_to_emit = emit_event_logic(event_buffer=event_buffer, event=event)
+                for emit_event in events_to_emit:
+                    yield emit_event
+
+    # Process ONLY completion cleanup events, not content from completion chunk
+    if completion_chunk:
+        completion_events = create_completion_events(
+            completion_chunk, event_buffer, message_started, message_id, thread_id, run_id, run_state=run_state
+        )
+        for event in completion_events:
+            events_to_emit = emit_event_logic(event_buffer=event_buffer, event=event)
+            for emit_event in events_to_emit:
+                yield emit_event
+
+    # Ensure completion events are always emitted even when stream ends naturally
+    if not stream_completed:
+        # Synthesize a completion event so the client always sees RUN_FINISHED
+        synthetic_completion = RunCompletedEvent()
+        completion_events = create_completion_events(
+            synthetic_completion, event_buffer, message_started, message_id, thread_id, run_id, run_state=run_state
+        )
+        for event in completion_events:
+            events_to_emit = emit_event_logic(event_buffer=event_buffer, event=event)
+            for emit_event in events_to_emit:
+                yield emit_event
